LIBDRUM-380. Move XMLUI local UMD customizations to edu.umd.lib package
Updated references to classes within other files.
Fixed package name.

https://issues.umd.edu/browse/LIBDRUM-380
