+++ dspace/modules/xmlui/src/main/java/edu/umd/lib/dspace/app/xmlui/aspect/administrative/etd_departments/DeleteETDDepartmentsConfirm.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2009 The University of Maryland. All Rights Reserved.
  *
  */
-package edu.umd.lib.dspace.app.xmlui.aspect.administrative.etd_departmens;
+package edu.umd.lib.dspace.app.xmlui.aspect.administrative.etd_departments;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
