+++ dspace/modules/xmlui/src/main/resources/aspects/Administrative/administrative.js
@@ -27,8 +27,8 @@ importClass(Packages.org.dspace.app.xmlui.utils.FlowscriptUtils);
 importClass(Packages.org.dspace.app.xmlui.utils.ContextUtil);
 importClass(Packages.org.dspace.app.xmlui.aspect.administrative.FlowEPersonUtils);
 importClass(Packages.org.dspace.app.xmlui.aspect.administrative.FlowGroupUtils);
-importClass(Packages.org.dspace.app.xmlui.aspect.administrative.FlowETDDepartmentUtils);
-importClass(Packages.org.dspace.app.xmlui.aspect.administrative.FlowUnitsUtils);
+importClass(Packages.edu.umd.lib.dspace.app.xmlui.aspect.administrative.FlowETDDepartmentUtils);
+importClass(Packages.edu.umd.lib.dspace.app.xmlui.aspect.administrative.FlowUnitsUtils);
 importClass(Packages.org.dspace.app.xmlui.aspect.administrative.FlowRegistryUtils);
 importClass(Packages.org.dspace.app.xmlui.aspect.administrative.FlowItemUtils);
 importClass(Packages.org.dspace.app.xmlui.aspect.administrative.FlowMapperUtils);
