+++ dspace/modules/xmlui/src/main/resources/aspects/Administrative/sitemap.xmap
@@ -29,13 +29,13 @@
             <map:transformer name="EditGroupForm" src="org.dspace.app.xmlui.aspect.administrative.group.EditGroupForm"/>
             <map:transformer name="DeleteGroupsConfirm" src="org.dspace.app.xmlui.aspect.administrative.group.DeleteGroupsConfirm"/>
 
-            <map:transformer name="ManageETDDepartmentsMain" src="org.dspace.app.xmlui.aspect.administrative.etd_departments.ManageETDDepartmentsMain"/>
-            <map:transformer name="EditETDDepartmentsForm" src="org.dspace.app.xmlui.aspect.administrative.etd_departments.EditETDDepartmentsForm"/>
-            <map:transformer name="DeleteETDDepartmentsConfirm" src="org.dspace.app.xmlui.aspect.administrative.etd_departments.DeleteETDDepartmentsConfirm"/>
+            <map:transformer name="ManageETDDepartmentsMain" src="edu.umd.lib.dspace.app.xmlui.aspect.administrative.etd_departments.ManageETDDepartmentsMain"/>
+            <map:transformer name="EditETDDepartmentsForm" src="edu.umd.lib.dspace.app.xmlui.aspect.administrative.etd_departments.EditETDDepartmentsForm"/>
+            <map:transformer name="DeleteETDDepartmentsConfirm" src="edu.umd.lib.dspace.app.xmlui.aspect.administrative.etd_departments.DeleteETDDepartmentsConfirm"/>
             
-            <map:transformer name="ManageUnitsMain" src="org.dspace.app.xmlui.aspect.administrative.units.ManageUnitsMain"/>
-            <map:transformer name="EditUnitsForm" src="org.dspace.app.xmlui.aspect.administrative.units.EditUnitsForm"/>
-            <map:transformer name="DeleteUnitsConfirm" src="org.dspace.app.xmlui.aspect.administrative.units.DeleteUnitsConfirm"/>
+            <map:transformer name="ManageUnitsMain" src="edu.umd.lib.dspace.app.xmlui.aspect.administrative.units.ManageUnitsMain"/>
+            <map:transformer name="EditUnitsForm" src="edu.umd.lib.dspace.app.xmlui.aspect.administrative.units.EditUnitsForm"/>
+            <map:transformer name="DeleteUnitsConfirm" src="edu.umd.lib.dspace.app.xmlui.aspect.administrative.units.DeleteUnitsConfirm"/>
 
             <map:transformer name="MetadataRegistryMain" src="org.dspace.app.xmlui.aspect.administrative.registries.MetadataRegistryMain"/>
             <map:transformer name="EditMetadataSchema" src="org.dspace.app.xmlui.aspect.administrative.registries.EditMetadataSchema"/>
